UniversityManagement/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2.2
django-bootstrap4==23.1
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
